CAMEL-1373: Use CamelCase for header/prop keys. Moved well knows key consts to Exchange. Polished code. Added surpruse for compiler unchecked warnings. Removed some left over generics on Route. Fixed a bug in some util class.
git-svn-id: https://svn.eu.apache.org/repos/asf/camel/trunk@747062 13f79535-47bb-0310-9956-ffa450edef68

+++ camel-core/src/main/java/org/apache/camel/processor/Delayer.java
@@ -71,7 +71,7 @@ public class Delayer extends DelayProcessorSupport {
         if (timeExpression != null) {
             Long longValue = timeExpression.evaluate(exchange, Long.class);
             if (longValue != null) {
-                time = longValue.longValue();
+                time = longValue;
             }
         }
         if (time <= 0) {
